Add admin category views and routes

+++ routes/web.php
@@ -40,11 +40,12 @@ Route::prefix('admin')/* ->middleware(['auth']) */->group(function () {
     Route::put('products/{id}/categories', [ProductController::class, 'syncCategories']);
 
     // categories
-    Route::get('categories', [CategoryController::class, 'index']);
-    Route::post('categories', [CategoryController::class, 'store']);
-    Route::get('categories/{id}', [CategoryController::class, 'show']);
-    Route::put('categories/{id}', [CategoryController::class, 'update']);
+    Route::get('categories', [CategoryController::class, 'index'])->name('admin.categories.index');
+    Route::get('categories/create', [CategoryController::class, 'create'])->name('admin.categories.create');
+    Route::post('categories', [CategoryController::class, 'store'])->name('admin.categories.store');
+    Route::get('categories/{id}/edit', [CategoryController::class, 'edit'])->name('admin.categories.edit');
+    Route::put('categories/{id}', [CategoryController::class, 'update'])->name('admin.categories.update');
     Route::patch('categories/{id}', [CategoryController::class, 'update']);
-    Route::delete('categories/{id}', [CategoryController::class, 'destroy']);
-    Route::get('categories/{id}/products', [CategoryController::class, 'products']);
+    Route::delete('categories/{id}', [CategoryController::class, 'destroy'])->name('admin.categories.destroy');
+    Route::get('categories/{id}/products', [CategoryController::class, 'products'])->name('admin.categories.products');
 });
